+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -5,6 +5,8 @@
 import FlutterMacOS
 import Foundation
 
+import secure_pinning_apple
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  SecurePinningPlugin.register(with: registry.registrar(forPlugin: "SecurePinningPlugin"))
 }